+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.24] - 2018-10-20
+### Added
+- PacketServer.Global.PsMovement transmit movement to all know players.
+- Now you can see another players in the world. :)
+
 ## [0.0.23] - 2018-10-20
 ### Added
 - Class Table Items added default value if not exist.